Move IllegalValueException to top level exception package
