+++ src/main/java/seedu/address/storage/JsonAdaptedTag.java
@@ -3,7 +3,7 @@ package seedu.address.storage;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonValue;
 
-import seedu.address.commons.exceptions.IllegalValueException;
+import seedu.address.exception.IllegalValueException;
 import seedu.address.model.tag.Tag;
 
 /**
