+++ src/test/java/seedu/address/storage/JsonSerializableAddressBookTest.java
@@ -8,8 +8,8 @@ import java.nio.file.Paths;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.commons.exceptions.IllegalValueException;
 import seedu.address.commons.util.JsonUtil;
+import seedu.address.exception.IllegalValueException;
 import seedu.address.model.AddressBook;
 import seedu.address.testutil.TypicalPersons;
